Move sub-classes of "GridEntry" class to their own files
